pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DataBUS"
version = "0.1.0"
description = "Upload dataset records to the Neotoma Database"
authors = [{name = "Socorro E Dominguez Vidana", email = "dominguezvid@uwisc.edu"}]
license = "MIT"
license-files = ["LICENSE.md"]
requires-python = "==3.11.10"
readme = {file = "README.md", content-type = "text/markdown"}
keywords = ["neotoma", "paleoecology", "interoperability", "postgres", "SQL"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Topic :: Database",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Software Development :: Object Brokering"
]
dependencies = [
    "numpy>1.25.0",
    "pandas==2.0.2",
    "psycopg2-binary>=2.9.10",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "pytz==2023.3",
    "PyYAML==5.3.1",
    "six==1.16.0",
    "tzdata==2023.3",
    "openpyxl",
    "requests",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
    "mkdocstrings[python-legacy]>=0.18",
]

[project.urls]
Issues = "https://github.com/NeotomaDB/DataBUS/issues"
Repository = "https://github.com/NeotomaDB/DataBUS.git"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"DataBUS" = ["sqlHelpers/upsert_site.sql",
      "sqlHelpers/insert_pb_model.sql",
      "sqlHelpers/insert_data_uncertainty.sql",
      "sqlHelpers/insert_taxa_row.sql",
      "sqlHelpers/upsert_collunit.sql",
      "sqlHelpers/insert_hiatus.sql",
      "sqlHelpers/insert_hiatuschronology.sql",
      "sqlHelpers/insert_uraniumseriesdata.sql",
      "sqlHelpers/insert_uthseries.sql",
      "sqlHelpers/insert_speleothem.sql",
      "sqlHelpers/insert_speleothem_cu.sql",
      "sqlHelpers/insert_entitycovers.sql",
      "sqlHelpers/insert_entitydripheight.sql",
      "sqlHelpers/insert_entitygeology.sql",
      "sqlHelpers/insert_entitylandusecover.sql",
      "sqlHelpers/insert_entityrelationship.sql",
      "sqlHelpers/insert_entitysamples.sql",
      "sqlHelpers/insert_entityvegetationcover.sql",
      "sqlHelpers/insert_externalspeleothem.sql",]

[project.optional-dependencies]
build = ["build", "twine"]
dev = ["ruff", "bumpver", "mypy", "pytest"]

[tool.ruff]
# Target Python 3.11
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
# Enable common rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (common in validators)
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
# Tests can use assertions and have unused imports for fixtures
"**/tests/*" = ["S101", "ARG", "F401"]
# __init__.py files often have unused imports for re-export
"__init__.py" = ["F401"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["src/DataBUS/tests", "src/DataBUS/neotomaHelpers/tests", "src/DataBUS/neotomaValidator/tests", "src/DataBUS/neotomaUploader/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
